requirements.txt
fastapi
uvicorn[standard]
scikit-learn==1.6.1
tensorflow==2.16.1
numpy==1.26.4
joblib==1.4.2
python-multipart
Pillow
openmeteo-requests
requests-cache
retry-requests
pydantic
xgboost
